feat🎨Added isSelected list of bool to change the color when the user click in the button.
