+++ lib/features/cart/ui/widgets/cart_items_list.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:shopping_app/core/theming/text_styles.dart';
 import 'package:shopping_app/core/widgets/spacing.dart';
+import 'package:shopping_app/features/cart/ui/widgets/counter_widget.dart';
 
 class CartItemsList extends StatefulWidget {
   const CartItemsList({super.key});
@@ -11,7 +12,7 @@ class CartItemsList extends StatefulWidget {
 }
 
 class _CartItemsListState extends State<CartItemsList> {
-  // List of image URLs and prices
+  // List of image URLs, prices, and titles
   final List<Map<String, String>> items = [
     {
       "image": "assets/images/Order_Items.png",
@@ -50,6 +51,16 @@ class _CartItemsListState extends State<CartItemsList> {
     },
   ];
 
+  // List to track whether each item is selected
+  late List<bool> isSelected;
+
+  @override
+  void initState() {
+    super.initState();
+    // Initialize the isSelected list with false for each item
+    isSelected = List<bool>.filled(items.length, false);
+  }
+
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
@@ -65,8 +76,10 @@ class _CartItemsListState extends State<CartItemsList> {
             width: double.infinity,
             decoration: BoxDecoration(
               border: Border.all(
-                color: const Color.fromRGBO(0, 0, 0, 490),
-                // Pink when selected, grey otherwise
+                // Change border color based on isSelected
+                color: isSelected[index]
+                    ? Colors.pink
+                    : const Color.fromRGBO(0, 0, 0, 490),
                 width: 2.0,
               ),
               borderRadius: BorderRadius.circular(13.0),
@@ -83,10 +96,7 @@ class _CartItemsListState extends State<CartItemsList> {
                   ),
                 ),
                 Padding(
-                  padding: const EdgeInsets.only(
-                    left: 8.0,
-                    top: 8,
-                  ),
+                  padding: const EdgeInsets.only(left: 8.0, top: 8),
                   child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
@@ -104,9 +114,17 @@ class _CartItemsListState extends State<CartItemsList> {
                           Padding(
                             padding: EdgeInsets.only(left: 16.w, right: 11.w),
                             child: InkWell(
-                              onTap: () {},
+                              onTap: () {
+                                setState(() {
+                                  // Toggle selection state
+                                  isSelected[index] = !isSelected[index];
+                                });
+                              },
                               child: Image.asset(
-                                'assets/images/icon_button_not_selected.png',
+                                // Change image based on isSelected
+                                isSelected[index]
+                                    ? 'assets/images/icon_button_is_selected.png'
+                                    : 'assets/images/icon_button_not_selected.png',
                                 fit: BoxFit.fill,
                                 width: 18.w,
                                 height: 18.h,
@@ -115,12 +133,10 @@ class _CartItemsListState extends State<CartItemsList> {
                           ),
                         ],
                       ),
-                      //  verticalSpace(4),
                       Text(
                         items[index]["price"]!,
                         style: TextStyles.font16BlackSemiBold,
                       ),
-                      // verticalSpace(6),
                       Row(
                         children: [
                           const CounterWidget(),
@@ -130,12 +146,12 @@ class _CartItemsListState extends State<CartItemsList> {
                               'Edit',
                               style: TextStyles.font12PinkRegular.copyWith(
                                 decoration: TextDecoration.underline,
-                                decorationColor: TextStyles.font12PinkRegular
-                                    .color, // Use the same color as the text
+                                decorationColor:
+                                    TextStyles.font12PinkRegular.color,
                               ),
                             ),
                             onPressed: () {},
-                          )
+                          ),
                         ],
                       ),
                     ],
@@ -150,73 +166,3 @@ class _CartItemsListState extends State<CartItemsList> {
   }
 }
 
-class CounterWidget extends StatefulWidget {
-  const CounterWidget({super.key});
-
-  @override
-  _CounterWidgetState createState() => _CounterWidgetState();
-}
-
-class _CounterWidgetState extends State<CounterWidget> {
-  int _counter = 1;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  void _decrementCounter() {
-    if (_counter > 0) {
-      setState(() {
-        _counter--;
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 30.h,
-      decoration: BoxDecoration(
-        color: Colors.pink,
-        borderRadius: BorderRadius.circular(30),
-      ),
-      child: Row(
-        //  mainAxisSize: MainAxisSize.min,
-        children: [
-          IconButton(
-            onPressed: _decrementCounter,
-            icon: const Icon(Icons.remove),
-            color: Colors.white,
-            iconSize: 20,
-            //  padding: const EdgeInsets.all(8),
-            constraints: const BoxConstraints(),
-            splashRadius: 20,
-            highlightColor: Colors.pink[200],
-            splashColor: Colors.pink[400],
-            tooltip: 'Decrement',
-            visualDensity: VisualDensity.adaptivePlatformDensity,
-          ),
-          Text(
-            '$_counter',
-            style: TextStyles.font14WhiteSemiBold,
-          ),
-          IconButton(
-            onPressed: _incrementCounter,
-            icon: const Icon(Icons.add),
-            color: Colors.white,
-            iconSize: 20,
-            // padding: const EdgeInsets.all(8),
-            constraints: const BoxConstraints(),
-            splashRadius: 20,
-            highlightColor: Colors.pink[200],
-            splashColor: Colors.pink[400],
-            tooltip: 'Increment',
-            visualDensity: VisualDensity.adaptivePlatformDensity,
-          ),
-        ],
-      ),
-    );
-  }
-}
